Refactor locale handling
During the crusade removing GlobalSettings from vf-qmllibs, it came back to
me. Locale handling has lots of room for enhancements:

* Keeping two central properies locale/localeName is dangerous:
  * locale was a function binding only. The risk of breaking this binding by
    setting it from JavaScript are high. To avoid introduce getter functions
  * In ZSpinBox were leftovers from my playgound [1]. Being lazy
    VirtualKeyboard was the keeper application's locale.
* In vf-declarative-gui ther were loads of different implementations dealing
  with locale - a patch for that is in the queue.

We replace ZeraHelpersConfig (yes it was just introduced) by ZeraLocale because
that is what this component is about. Since they are dealing with locales, the
functions strToCLocale and strToLocal were moved to ZeraLocale.

[1] https://github.com/schnitzeltony/qtvirtualkeyboard-demo

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>