EthernetSettings / WifiSettings: Rework Client/Hotspot ComboBox fill/select
Move translation and plausibility issues at one place: Our backend. That is the
object that gives us the interface to underlying networkmanger-code and knows
all the plausibilites best. Same for translation: Keep seperate models for
backend and display once and use display model only for populating our
comboboxes. That avoids surprises once we start with translation.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>