Show connection/disconnecting status by setting switch to pos 0.5.
In future we want to remove the nm-applet. Until now this application does not show
when etablishing a connection is in progress.

Pressing the button will set the pos to 0.5. If the status is connected the pos will be 1
or if status is disconnected then pos is 0. To make that work it was also
necessary to move connected flag setter from connectionActivated to connection
statusChangedReason slot.

Signed-off-by: bhamacher <b.hamacher@zera.de>