Differentiate between new and old connections settingsInterfaces.
When we try to saveandExit a connection, that already exis the old
connection is deleted and replaced with the new one.

This is preperation to use this function in case the authentication failed as well.
The old connection will be replaced with a the new one using the new password.

Signed-off-by: bhamacher <b.hamacher@zera.de>